requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
google-cloud-bigquery
google-cloud-storage
google-cloud-firestore
google-adk[all]
pandas
openpyxl
python-dotenv
python-multipart
PyJWT
cryptography
passlib[bcrypt]

# Testing
pytest
pytest-mock
pytest-asyncio
requests

# BigQuery
db-dtypes

#Cloud Run
gunicorn; sys_platform != 'win32'
uvloop; sys_platform != 'win32'
tenacity